requirements.txt
asgiref==3.8.1
Django==5.0.3
django-environ==0.11.2
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
python-telegram-bot==21.0.1
factory-boy==3.3.0
freezegun==1.5.1